further updated files, and pairwise file
Cleaned up a bit the files, and adding the data_pairwise.csv. the file contains the pairwise distance and is in the form MELT, instead of having it in a matrix. For each pair i annotate also the geo distance in km, and the approximate divergence time calculated from effective population size and linear FST. the pairwise statistics are also explained in the file variables.csv, together with the statistics from data.csv